Step Two: Insecure Direct Object References
Just because a REST client has the primary key to an entity, it doesn't
mean that they are authorized to read it. You can use @PostAuthorize to
check whether the owner of the resource belongs to the logged in user.

If there is a circumstance when we can't filter by logged-in user at the
query level, we can use the @PostFilter annotation to do it at the
method level. Note that modifying @Query expression is preferred,
though, as it will scale much better.

Adding to a query's where clause is a nice, scalable way to filter
results down to only those owned by the user.

Note that @PostAuthorize is still needed, if you want to throw an 403
instead of a 404, since once you add the user to the query, it might
return no results at all.